+++ ContentPatcher/release-notes.md
@@ -4,6 +4,7 @@
 ## Upcoming release
 * Fixed `Day` token not allowing zero values.
 * Fixed dynamic tokens validated before they're ready.
+* Fixed custom tokens called with non-ready inputs in some cases.
 
 ## 1.9.1
 Released 12 June 2019.